pyproject.toml
[tool.poetry]
name = "String Toolkit"
version = "2.0.0"
description = "Run different sets of string manipulation tools"
documentation = "https://github.com/fedecalendino/alfred-string-toolkit/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-string-toolkit"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
alfred-pyflow = "^0.4.5"
lorem-text = "^2.1"
stringcase = "^1.2.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
